Introduce Optimizer Interface and adapt Logging to it
This introduces a common Interface between Optimizer and IterativeOptimizerand adapts
the tracing to it. It also renames the Optimizer to TopDownOptimizer because this
is what conceptionally and how it's called in literature.
